Fix minor typo in example toml configs (#364)
As I was familiarizing myself with this repo and getting it to run
locally I noticed a minor typo in the github section for both the
rfd-api and rfd-processor example config files.

The comment for the `default branch` should read:

"Branch to use as the default branch of the repository"

instead of

"Branch to use as the default branch of ther repository"

Just changed "ther" to "the".